4.1.1. Добавлено наследование простым компонентам
- Создан компонент AbstractView;
- Простые компоненты унаследовали AbstractView;
