+++ src/view/events-list-view.js
@@ -1,22 +1,9 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createEventsListTemplate = () => ('<ul class="trip-events__list"></ul>');
 
-export default class EventsListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
+export default class EventsListView extends AbstractView {
   get template() {
     return createEventsListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
